Fixed pixel selection accuracy

+++ HilbertCurve.java
@@ -63,8 +63,8 @@ public class HilbertCurve {
 		o.setVisible(true);		// initialize everything
 	}
 
-	public void hilbert(int x, int y, int xi, int xj, int yi, int yj, int n) {
-		if (n <= 0) h.addPoint(x + (xi + yi) / 2, y + (xj + yj) / 2);
+	public void hilbert(float x, float y, float xi, float xj, float yi, float yj, int n) {
+		if (n <= 0) h.addPoint((int) (x + (xi + yi) / 2), (int) (y + (xj + yj) / 2));
 		else {
 			hilbert(x, y, yi / 2, yj / 2, xi / 2, xj / 2, n - 1);
 			hilbert(x + xi / 2, y + xj / 2, xi / 2, xj / 2, yi / 2, yj / 2, n - 1);
